data/requirements.txt
streamlit
pandas
openai
google-generativeai
tqdm
scikit-learn
sentence-transformers
faiss-cpu
